+++ packages/tools/src/utilities/contours/AnnotationToPointData.ts
