+++ packages/tools/src/utilities/contours/detectContourHoles.ts
