+++ packages/tools/src/utilities/contours/mergePoints.ts
