+++ packages/tools/src/utilities/index.ts
@@ -26,7 +26,7 @@ import { pointToString } from './pointToString';
 import annotationFrameRange from './annotationFrameRange';
 
 // name spaces
-import * as rtstruct from './rtstruct';
+import * as contours from './contours';
 import * as segmentation from './segmentation';
 import * as drawing from './drawing';
 import * as math from './math';
@@ -64,7 +64,7 @@ export {
   getCalibratedAreaUnits,
   getCalibratedScale,
   segmentation,
-  rtstruct,
+  contours,
   triggerAnnotationRenderForViewportIds,
   triggerAnnotationRender,
   pointInShapeCallback,
